+++ tests/conftest.py
@@ -385,3 +385,8 @@ def application_user(db, oauth_client) -> FidesUser:
     oauth_client.save(db=db)
     yield user
     user.delete(db=db)
+
+
+@pytest.fixture(autouse=True)
+def clear_get_config_cache() -> None:
+    get_config.cache_clear()