+++ CHANGELOG.md
@@ -36,7 +36,8 @@ The types of changes are:
 * Added user auth tables [915](https://github.com/ethyca/fides/pull/915)
 * Standardized API error parsing under `~/types/errors`
 * Added taxonomy page to UI [#902](https://github.com/ethyca/fides/pull/902)
-* Add user auth routes [https://github.com/ethyca/fides/pull/929](929)
+* Add lru cache to get_config [927](https://github.com/ethyca/fides/pull/927)
+* Add user auth routes [929](https://github.com/ethyca/fides/pull/929)
 
 ### Changed